Improve RecoveryMiddleware() to handle ErrAbortHandler panic
ReverseProxy() will raise this panic when it fails to copy response to
the client (e.g., client prematurely closes the connection).  Propagate
this panic and let http.Server handle it, which basically ignores it and
aborts the connection.

Meanwhile, add a unit test for this case.

Reference: doc of http.Handler:
```
If ServeHTTP panics, the server (the caller of ServeHTTP) assumes that
the effect of the panic was isolated to the active request. It recovers
the panic, logs a stack trace to the server error log, and either closes
the network connection or sends an HTTP/2 RST_STREAM, depending on the
HTTP protocol. To abort a handler so the client sees an interrupted
response but the server doesn't log an error, panic with the value
ErrAbortHandler.
```

Credits:
- https://github.com/golang/go/issues/28239
- https://github.com/go-chi/chi/issues/588